src/auth/requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
parso==0.8.3
pycparser==2.21
PyJWT==2.6.0
Werkzeug==2.2.2
zipp==3.10.0
